First attempt at retry logic
Add retry logic to getUsersEmail. This retry logic only looks for cases where GMail is quota-limiting us. It could be expanded to retry in other cases.

I am not happy with this logic. It is too complicated and would have to be duplicated to each GMail API call. I would rather refactor the code to have a generic retry function that can operate on any GMail API call, but so far the code for that has eluded me.